Split objects library into objects, exceptions and objects_internal libraries.
(#86c778p0x)
