+++ README.md
@@ -1,6 +1,6 @@
 # Norman Objects
 
-Norman Objects is a shared library that defines the core request and response classes used across the Norman SDK and internal services.
+Norman Objects is a shared library that defines the core request and response classes used across the Norman SDK.
 
 The library includes classes, enums, and other structured objects that establish clear, consistent, type-safe contracts between backend services and SDK clients.
