+++ norman_objects/services/file_pull/download/tracked_asset_download.py
@@ -8,7 +8,7 @@ from norman_objects.shared.context.norman_access_context import NormanAccessCont
 from norman_objects.shared.files.file_properties import FileProperties
 from norman_objects.shared.messages.asset_message import AssetMessage
 from norman_objects.shared.messages.entity_type import EntityType
-from norman_objects.shared.models.model_asset import ModelAsset
+from norman_objects.shared.assets.model_asset import ModelAsset
 from norman_objects.shared.status_flags.status_flag import StatusFlag
 from norman_objects.shared.status_flags.status_flag_name import StatusFlagName
 from norman_objects.shared.status_flags.status_flag_value import StatusFlagValue
