--- norman_objects/shared/accounts/account_profile.py
@@ -1,7 +1,0 @@
-from typing import Optional
-
-from norman_objects.shared.accounts.account import Account
-
-
-class AccountProfile(Account):
-    email: Optional[str] = None
