+++ norman_objects/shared/assets/model_asset.py
