--- norman_objects/shared/authentication/argon2_algorithm_enum.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-class Argon2AlgorithmEnum(str, Enum):
-    ID = "ID"
-    I = "I"
-    D = "D"
