--- norman_objects/shared/context/noop_context_manager.py
@@ -1,9 +1,0 @@
-class NoopContextManager:
-    def __init__(self):
-        pass
-
-    def __enter__(self):
-        pass
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        pass
