--- norman_objects/shared/databases/lock_type.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class LockType(str, Enum):
-    Read = "Read"
-    Write = "Write"
