--- norman_objects/shared/exceptions/server_exception.py
@@ -1,21 +1,0 @@
-from norman_objects.shared.exceptions.norman_exception import NormanException
-
-
-class ServerException(NormanException):
-    status_code: int = 500
-    error_type: str = "server"
-
-    def __init__(
-            self,
-            message: str,
-            cause: str,
-            suggestions: list[str]
-    ):
-
-        super().__init__(
-            status_code=self.status_code,
-            error_type=self.error_type,
-            message=message,
-            cause=cause,
-            suggestions=suggestions,
-        )
