--- norman_objects/shared/functions/function_state.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class FunctionState(str, Enum):
-    Active = "Active"
-    Pending = "Pending"
-    Inactive = "Inactive"
-    Failed = "Failed"
-    NotProvided = "NotProvided"
