--- norman_objects/shared/messages/entity_type.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class EntityType(str, Enum):
-    Model = "Model"
-    Asset = "Asset"
-    Invocation = "Invocation"
-    Input = "Input"
-    Output = "Output"
