--- norman_objects/shared/messages/norman_message_union.py
@@ -1,20 +1,0 @@
-from typing import Union, Annotated
-
-from pydantic import Field
-
-from norman_objects.shared.messages.asset_message import AssetMessage
-from norman_objects.shared.messages.input_message import InputMessage
-from norman_objects.shared.messages.invocation_message import InvocationMessage
-from norman_objects.shared.messages.model_message import ModelMessage
-from norman_objects.shared.messages.output_message import OutputMessage
-
-NormanMessageUnion = Annotated[
-    Union[
-        ModelMessage,
-        AssetMessage,
-        InvocationMessage,
-        InputMessage,
-        OutputMessage
-    ],
-    Field(discriminator="entity_type")
-]
