+++ norman_objects/shared/models/model_base.py
@@ -4,7 +4,7 @@ from pydantic import Field
 
 from norman_objects.norman_base_model import NormanBaseModel
 from norman_objects.shared.date.normalized_datetime import NormalizedDateTime
-from norman_objects.shared.models.aggregate_tag import AggregateTag
+from norman_objects.shared.tags.aggregate_tag import AggregateTag
 
 
 class ModelBase(NormanBaseModel):
