+++ norman_objects/shared/models/model_preview.py
@@ -3,7 +3,7 @@ from datetime import datetime, timezone
 from pydantic import Field
 
 from norman_objects.shared.date.normalized_datetime import NormalizedDateTime
-from norman_objects.shared.models.aggregate_tag import AggregateTag
+from norman_objects.shared.tags.aggregate_tag import AggregateTag
 from norman_objects.shared.models.model_base import ModelBase
 from norman_objects.shared.models.model_version_preview import ModelVersionPreview
