--- norman_objects/shared/provisioning/provisioned_instance_status.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class ProvisionedInstanceStatus(str, Enum):
-    Pending = "Pending"
-    Running = "Running"
-    ShuttingDown = "ShuttingDown"
-    Terminated = "Terminated"
