+++ norman_objects/shared/versions/http_request_type.py
