+++ norman_objects/shared/versions/model_hosting_location.py
