+++ norman_objects/shared/versions/model_version.py
@@ -6,7 +6,7 @@ from pydantic import Field
 from norman_objects.shared.date.normalized_datetime import NormalizedDateTime
 from norman_objects.shared.model_signatures.model_signature import ModelSignature
 from norman_objects.shared.models.http_request_type import HttpRequestType
-from norman_objects.shared.models.model_asset import ModelAsset
+from norman_objects.shared.assets.model_asset import ModelAsset
 from norman_objects.shared.models.model_build_status import ModelBuildStatus
 from norman_objects.shared.models.model_hosting_location import ModelHostingLocation
 from norman_objects.shared.models.model_type import ModelType
