+++ norman_objects/shared/versions/model_version_preview.py
@@ -1,5 +1,5 @@
 from norman_objects.norman_base_model import NormanBaseModel
-from norman_objects.shared.models.model_asset import ModelAsset
+from norman_objects.shared.assets.model_asset import ModelAsset
 from norman_objects.shared.models.model_build_status import ModelBuildStatus
