--- requirements.txt.template
@@ -1,3 +1,0 @@
-build==0.10.0
-pydantic==2.12.4
-setuptools==67.6.0
